refactor(canvas-chat): extract serializeEditable to its own module
The earlier paste-newline fix pushed mentions.ts past the 500-line file-size
governance gate. Move serializeEditable (and its block-tag table) into a
dedicated module and re-export it from mentions.ts, so existing importers are
unaffected and mentions.ts drops back under the limit. No behavior change.
